backend/orchestrator/pyproject.toml
[tool.poetry]
name = "tender-orchestrator"
version = "0.1.0"
description = "Tender Analysis System Backend"
authors = ["Your Name <serigne.faye@alggroup.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.5"
uvicorn = {extras = ["standard"], version = "^0.32.1"}
pydantic = "^2.10.3"
pydantic-settings = "^2.6.1"
motor = "^3.6.0"
beanie = "^1.27.0"
pymongo = "^4.10.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.17"
python-dotenv = "^1.0.1"
httpx = "^0.28.1"
openai = "^1.57.0"
langchain = "^0.3.12"
langchain-openai = "^0.2.11"
langchain-community = "^0.3.11"
pandas = "^2.2.3"
numpy = "^2.2.6"
redis = "^5.2.1"
aiofiles = "^24.1.0"
tenacity = "^9.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
ruff = "^0.8.4"
mypy = "^1.13.0"
ipython = "^8.30.0"
black = "^24.10.0"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict_optional = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --cov=. --cov-report=term-missing"